+++ web/src/components/core/TableKelas.vue
@@ -20,6 +20,7 @@ export default defineComponent({
       createdItem: { ...defaultItem },
       items: [],
       showModal: false,
+      loading: false,
     };
   },
 
@@ -38,7 +39,16 @@ export default defineComponent({
   },
 
   methods: {
+    /**
+     * Fetches kelas data from the API and updates component state.
+     *
+     * Makes authenticated request to API endpoint using stored JWT token.
+     * Handles loading state and errors.
+     * Maps API response data to format needed for component state.
+     */
     async fetchData() {
+      this.loading = true;
+
       try {
         const jwtToken = localStorage.getItem("jwtToken");
 
@@ -65,8 +75,19 @@ export default defineComponent({
         }));
       } catch (error) {
         console.error("Error fetching data:", error);
+      } finally {
+        this.loading = false;
       }
     },
+    /**
+     * Adds a new kelas item by making a POST request to the API.
+     * - Gets the JWT token from localStorage.
+     * - Makes a POST request to the API with the new item data.
+     * - Handles API response and adds new item to component's items array.
+     * - Resets the component's createdItem data.
+     * - Refetches latest data from API after adding new item.
+     * - Handles any errors from the API request.
+     */
 
     async addNewItem() {
       try {
@@ -96,12 +117,24 @@ export default defineComponent({
 
         this.items.push(newKelas);
         this.resetCreatedItem();
-        this.fetchData();
+        await this.fetchData();
       } catch (error) {
         console.error("Error adding new item:", error);
       }
     },
 
+    /**
+     * Edit an existing kelas item by making a PUT request.
+     *
+     * Fetch the JWT token from localStorage to authenticate the request.
+     * Construct the edited kelas data object to send in the request body.
+     * Make the PUT request to the kelas API endpoint with the item ID.
+     * Update the item in the local items array with the response data.
+     * Reset the editedItem state.
+     * Refetch the kelas data.
+     *
+     * @throws {Error} If the request fails.
+     */
     async editItem() {
       try {
         const jwtToken = localStorage.getItem("jwtToken");
@@ -134,37 +167,53 @@ export default defineComponent({
         };
 
         this.resetEditedItem();
-        this.fetchData();
+        await this.fetchData();
       } catch (error) {
         console.error("Error editing item:", error);
       }
     },
 
+    /**
+     * Deletes a kelas by ID.
+     *
+     * Prompts user to confirm deletion.
+     * Checks for JWT token and adds authorization header.
+     * Makes DELETE request to API with kelas ID.
+     * Removes kelas from items array.
+     * Refetches data from API after deletion.
+     * Shows alert on success.
+     * Logs error on failure.
+     */
     async deleteItemById(id) {
-      try {
-        const jwtToken = localStorage.getItem("jwtToken");
+      if (window.confirm("Are you sure you want to delete this item?")) {
+        try {
+          const jwtToken = localStorage.getItem("jwtToken");
 
-        if (!jwtToken) {
-          console.error("JWT token not available");
-          // Handle the case where the token is not available (e.g., redirect to login)
-          return;
+          if (!jwtToken) {
+            console.error("JWT token not available");
+            // Handle the case where the token is not available (e.g., redirect to login)
+            return;
+          }
+
+          // Add headers with the authentication token
+          const headers = {
+            Authorization: `Bearer ${jwtToken}`,
+          };
+
+          // Make the DELETE request with headers
+          await axios.delete(
+            `http://localhost:3000/api/kelas/${this.items[id].id}`,
+            { headers },
+          );
+
+          // Remove the item from the array
+          this.items.splice(id, 1);
+          await this.fetchData();
+          // Optionally, you can show a success message
+          alert("Item deleted successfully");
+        } catch (error) {
+          console.error("Error deleting item:", error);
         }
-
-        // Add headers with the authentication token
-        const headers = {
-          Authorization: `Bearer ${jwtToken}`,
-        };
-
-        // Make the DELETE request with headers
-        await axios.delete(
-          `http://localhost:3000/api/kelas/${this.items[id].id}`,
-          { headers },
-        );
-
-        // Remove the item from the array
-        this.items.splice(id, 1);
-      } catch (error) {
-        console.error("Error deleting item:", error);
       }
     },
 
@@ -214,7 +263,7 @@ export default defineComponent({
     </va-button-group>
   </div>
   <div>
-    <va-data-table :items="items" :columns="columns" striped>
+    <va-data-table :items="items" :columns="columns" :loading="loading" striped>
       <template #cell(actions)="{ rowIndex }">
         <div class="action-buttons">
           <!--<va-button
